# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: Bazar AI API
  version: 0.2.0
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  description: |
    ## Контракт
    - Тело: один JSON-объект; лишние поля отклоняются (`DisallowUnknownFields`).
    - Ошибки: обёртка `{"error":{"code","message","request_id"?}}` ([ErrorEnvelope](#/components/schemas/ErrorEnvelope)).
    - Заголовки: ответ содержит `X-Request-Id`; при необходимости передавайте свой. Для refresh — `X-Device-Id` (опционально, должен совпадать при ротации).
    - Публичный заказ: заголовок `Idempotency-Key` (опционально) пробрасывается в заказ.
    - Лимит тела JSON: переменная окружения `API_MAX_BODY_BYTES` (по умолчанию 1 MiB). `multipart/*` на этом middleware не режется — отдельные лимиты в обработчиках (например 5 MiB для картинок).
    - Rate limit: **120** запросов / минуту с одного IP (код `rate_limited`, HTTP **429**).
    - CORS: если `ALLOWED_ORIGINS` не `*`, запросы с неразрешённого `Origin` → **403** `forbidden`.
...
servers:
  - url: http://localhost:8080
    description: Локальная разработка

# По умолчанию все операции требуют Bearer JWT; публичные переопределяют через security: []
security:
  - bearerAuth: []
paths:
...
  /api/v1/auth/register:
    post:
      security: []
      operationId: authRegister
      tags: [Auth]
      summary: Register seller
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRegisterRequest"
            example:
              email: seller@example.com
              password: secret42x
      responses:
        "200":
          description: Пользователь (без пароля) и пара токенов
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokenWithUser"
        "400":
          $ref: "#/components/responses/ValidationError"
        "409":
          $ref: "#/components/responses/Conflict"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/auth/login:
    post:
      security: []
      operationId: authLogin
      tags: [Auth]
      summary: Login seller
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthLoginRequest"
      responses:
        "200":
          description: Пользователь и пара токенов
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokenWithUser"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/auth/refresh:
    post:
      security: []
      operationId: authRefresh
      tags: [Auth]
      summary: Refresh access token
      description: Опциональный заголовок `X-Device-Id` должен совпадать с сохранённым для refresh-токена.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRefreshRequest"
      responses:
        "200":
          description: Новая пара токенов (без объекта `user`)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokenPair"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/auth/logout:
    post:
      security: []
      operationId: authLogout
      tags: [Auth]
      summary: Revoke refresh token
      description: |
        `all: true` отзывает все сессии, но тогда нужен **Bearer** в заголовке (как у `logout-all`).
        Иначе укажите `refresh_token`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthLogoutRequest"
      responses:
        "200":
          description: Сессия завершена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoggedOutStatus"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/auth/logout-all:
    post:
      operationId: authLogoutAll
      tags: [Auth]
      summary: Revoke all refresh tokens for current user
      responses:
        "200":
          description: Все сессии отозваны
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoggedOutStatus"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/auth/me:
    get:
      operationId: authMe
      tags: [Auth]
      summary: Current user profile
      responses:
        "200":
          description: Профиль пользователя
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserWrapped"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/onboarding/complete:
    post:
      operationId: onboardingComplete
      tags: [Onboarding]
      summary: Complete onboarding and create public store
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnboardingCompleteRequest"
      responses:
        "201":
          description: Магазин и публичный URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnboardingCompleteResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "402":
          $ref: "#/components/responses/QuotaExceeded"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/stores:
    post:
      operationId: createStore
      tags: [Stores]
      summary: Create store
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoreCreateRequest"
      responses:
        "201":
          description: Магазин создан
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/stores/me:
    get:
      operationId: listMyStores
      tags: [Stores]
      summary: Seller stores (paged)
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Пагинированный список магазинов
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListResponseStores"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/stores/{id}:
    get:
      operationId: getStoreById
      tags: [Stores]
      summary: Store by id
      parameters:
        - $ref: "#/components/parameters/StoreId"
      responses:
        "200":
          description: Объект магазина
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/stores/{storeID}/products:
    get:
      operationId: listStoreProducts
      summary: List products by store
      tags: [Products]
      parameters:
        - $ref: "#/components/parameters/StoreID"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Пагинированный список товаров
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListResponseProducts"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      operationId: createStoreProduct
      summary: Create product in store
      tags: [Products]
      parameters:
        - $ref: "#/components/parameters/StoreID"
      requestBody:
        required: true
        description: Поле `store_id` в теле перезаписывается из path.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductWriteRequest"
      responses:
        "201":
          description: Товар создан
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "402":
          $ref: "#/components/responses/QuotaExceeded"
        "403":
          $ref: "#/components/responses/Forbidden"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/products/{id}:
    get:
      operationId: getProductById
      summary: Get product by id
      tags: [Products]
      parameters:
        - $ref: "#/components/parameters/ProductId"
      responses:
        "200":
          description: Объект товара
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
    patch:
      operationId: updateProduct
      summary: Update product
      tags: [Products]
      parameters:
        - $ref: "#/components/parameters/ProductId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductWriteRequest"
      responses:
        "200":
          description: Обновлённый товар
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      operationId: deleteProduct
      summary: Delete product
      tags: [Products]
      parameters:
        - $ref: "#/components/parameters/ProductId"
      responses:
        "200":
          description: Товар удалён
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDeletedStatus"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/products/{id}/images:
    post:
      operationId: uploadProductImage
      tags: [Products]
      summary: Upload product image (multipart, поле `image`)
      description: Лимит тела ~5 MiB; глобальный JSON MaxBytes на этом маршруте не применяется.
      parameters:
        - $ref: "#/components/parameters/ProductId"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [image]
              properties:
                image:
                  type: string
                  format: binary
      responses:
        "201":
          description: URL загруженного изображения
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImageUploadResult"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/uploads/images:
    post:
      operationId: uploadImageAlias
      tags: [Products]
      summary: Upload image (multipart, alias)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [image]
              properties:
                image:
                  type: string
                  format: binary
      responses:
        "201":
          description: URL загруженного изображения
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImageUploadResult"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/public/stores/{slug}:
    get:
      security: []
      operationId: getPublicStorefront
      tags: [Public]
      summary: Public storefront data
      parameters:
        - $ref: "#/components/parameters/Slug"
      responses:
        "200":
          description: Витрина, товары и SEO-мета
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicStorefront"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/public/stores/{slug}/orders:
    post:
      security: []
      operationId: createPublicOrder
      tags: [Public]
      summary: Guest checkout
      description: Заголовок `Idempotency-Key` необязателен; сохраняется в заказе.
      parameters:
        - $ref: "#/components/parameters/Slug"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderCreateRequest"
      responses:
        "201":
          description: Созданный заказ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/stores/{storeID}/orders:
    get:
      operationId: listStoreOrders
      summary: List orders for store
      tags: [Orders]
      parameters:
        - $ref: "#/components/parameters/StoreID"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Пагинированный список заказов
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListResponseOrders"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/orders/{id}/notifications:
    get:
      operationId: listOrderNotifications
      summary: Notification jobs for order
      tags: [Orders]
      parameters:
        - $ref: "#/components/parameters/OrderId"
      responses:
        "200":
          description: Очередь уведомлений по заказу
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NotificationJob"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/stores/{storeID}/telegram/connect-code:
    post:
      operationId: createTelegramConnectCode
      summary: Issue Telegram pairing code
      tags: [Telegram]
      parameters:
        - $ref: "#/components/parameters/StoreID"
      responses:
        "201":
          description: Код и инструкция для бота
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TelegramConnectCodeResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/stores/{storeID}/telegram/status:
    get:
      operationId: getTelegramStatus
      summary: Telegram connection status
      tags: [Telegram]
      parameters:
        - $ref: "#/components/parameters/StoreID"
      responses:
        "200":
          description: >-
            Подключено — объект TelegramConnection; иначе только status not_connected.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/TelegramConnection"
                  - $ref: "#/components/schemas/TelegramNotConnected"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/stores/{storeID}/telegram/test:
    post:
      operationId: sendTelegramTestMessage
      summary: Send test order to Telegram
      tags: [Telegram]
      parameters:
        - $ref: "#/components/parameters/StoreID"
      responses:
        "200":
          description: Тестовое сообщение отправлено
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TelegramTestSent"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "502":
          $ref: "#/components/responses/BadGateway"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/telegram/webhook:
    post:
      security: []
      operationId: telegramWebhook
      tags: [Telegram]
      summary: Telegram Bot API webhook
      description: Подмножество Update Bot API; успех — `connected` или `ignored`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TelegramWebhookUpdate"
      responses:
        "200":
          description: Успешная обработка
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TelegramWebhookAck"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/ai/generate-store:
    post:
      operationId: aiGenerateStore
      tags: [AI]
      summary: AI generate store content
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AIGenerateRequest"
      responses:
        "200":
          description: Сохранённая запись генерации
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AIGeneration"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "402":
          $ref: "#/components/responses/QuotaExceeded"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "502":
          $ref: "#/components/responses/BadGateway"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/ai/generate-product:
    post:
      operationId: aiGenerateProduct
      summary: AI generate product copy
      tags: [AI]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AIGenerateRequest"
      responses:
        "200":
          description: Сохранённая запись генерации
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AIGeneration"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "402":
          $ref: "#/components/responses/QuotaExceeded"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "502":
          $ref: "#/components/responses/BadGateway"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/ai/generate-seo:
    post:
      operationId: aiGenerateSeo
      summary: AI generate SEO text
      tags: [AI]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AIGenerateRequest"
      responses:
        "200":
          description: Сохранённая запись генерации
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AIGeneration"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "402":
          $ref: "#/components/responses/QuotaExceeded"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "502":
          $ref: "#/components/responses/BadGateway"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/dashboard/leads:
    get:
      security: []
      operationId: sprintDashboardLeads
      tags: [Dashboard]
      summary: Dashboard leads
      responses:
        "200":
          description: Лиды по магазинам владельца
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardLeadsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    AuthRegisterRequest:
      type: object
      description: Нужен хотя бы один из `email` / `phone`; пароль ≥8 символов, буквы и цифры.
      properties:
        phone:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
...